Add Seq framework, fixed /stuck output, adjust movement to be bearable
