+++ Projects/CoX/Common/GameData/CMakeLists.txt
@@ -50,8 +50,9 @@ set(GAMEDATA_SRC
     #def_serializers.cpp
     map_serializers.cpp
     map_serializers.h
-    #seq_serializers.cpp
-    #seq_serializers.h
+    seq_definitions.h
+    seq_serializers.cpp
+    seq_serializers.h
     shop_serializers.cpp
     shop_serializers.h
     trick_serializers.cpp
