+++ Projects/CoX/Servers/MapServer/SlashCommand.cpp
@@ -867,10 +867,10 @@ void cmdHandler_Stuck(const QString &cmd, MapClientSession &sess)
     // TODO: Implement true move-to-safe-location-nearby logic
     forcePosition(*sess.m_ent,sess.m_current_map->closest_safe_location(sess.m_ent->m_entity_data.m_pos));
 
-    QString msg = QString("Resetting location to default spawn (%1,%2,%3)")
-                      .arg(sess.m_ent->m_entity_data.m_pos.x)
-                      .arg(sess.m_ent->m_entity_data.m_pos.y)
-                      .arg(sess.m_ent->m_entity_data.m_pos.z);
+    QString msg = QString("Resetting location to default spawn <%1, %2, %3>")
+                      .arg(sess.m_ent->m_entity_data.m_pos.x, 0, 'f', 1)
+                      .arg(sess.m_ent->m_entity_data.m_pos.y, 0, 'f', 1)
+                      .arg(sess.m_ent->m_entity_data.m_pos.z, 0, 'f', 1);
     qCDebug(logSlashCommand) << cmd << ":" << msg;
     sendInfoMessage(MessageChannel::SERVER, msg, &sess);
 }
